run.c - Enhance arg parse : optional parameters
It would be ideal to specify parameters without having to set all. One could use getopt for that but it is not portable. So we roll our own minimal one.

Usage:

```bash
run -c <checkpoint_file> -t [temperature] -s [steps] -p [prompt]
```

Minimum:

```bash
run -c <checkpoint_file>
```

The following are optional:

-t [temperature] -s [steps] -p [prompt]

if they are omitted, default values are used.

The order does not matter.

To keep it simple without needing for extended checks, if the parameter to the last option is missing, the current default value is used instead.

Updated README.MD to reflect changes.