Enhance arg parse : checkpoint 1st, -t -s -p optional
It would be ideal to specify parameters without having to set all. One could use getopt for that but it is not portable. So we roll our own minimal one.

Usage:

```bash
run <checkpoint_file> -t [temperature] -s [steps] -p [prompt]
```

Minimum:

```bash
run <checkpoint_file>
```

The following are optional:

-t [temperature] -s [steps] -p [prompt]

if they are omitted, default values are used.

The order of -t -s -p does not matter but the checkpoint_file has to be the first parameter.

To keep it simple without needing for extended checks, if the parameter to the last option is missing, the current default value is used instead.

Updated README.MD to reflect changes.

Incorporated requirement from: https://github.com/karpathy/llama2.c/pull/218#issuecomment-1662596666