Implement token healing
Fix the so called "mid-token" issue by stripping the last token from the prompt and then sampling the first processed token after the prompt among the ones that start with the last token's string.

We also got a new feature for free. We can now explicitly set last word boundary in the prompt, i.e. signal if the last word of the prompt should be considered a full word or just a partial word.

For instance, in this case:

```
./run model.bin 0 5 "On"
<s>
One day, a little
```

Last word of the prompt is expanded On -> One by the model. Instead if we want the sentence to specifically start with the word "On" we can now simply submit:

```
./run model.bin 0 5 "On "
<s>
On a hot summer day
```